app/requirements.txt
fastapi
uvicorn
gunicorn
pydantic
sqlalchemy
python-dotenv
pyodbc
bcrypt
pyjwt
pymssql